requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.11.1
httplib2==0.9.2
google-cloud==0.19.0
gunicorn==19.6.0
python-dateutil
requests
# oauth2client==4.0.0
oauth2client==2.0.1
requests.toolbelt
pyyaml

